+++ src/main/php/web/io/Incomplete.class.php
@@ -51,17 +51,6 @@ class Incomplete extends Part implements InputStream {
   public function bytes() { throw $this->notSupported(); }
 
   /**
-   * Transfers this stream to a given target.
-   *
-   * @deprecated Use `yield from $stream->transmit(...)` instead!
-   * @param  io.Path|io.Folder|io.streams.OutputStream|string $target
-   * @return int Number of bytes written
-   * @throws lang.IllegalArgumentException if filename is invalid
-   * @throws io.IOException
-   */
-  public function transfer($target) { throw $this->notSupported(); }
-
-  /**
    * Transmits this stream to a given target.
    *
    * @param  io.Path|io.Folder|io.streams.OutputStream|string $target