+++ src/test/php/web/unittest/ResponseTest.class.php
@@ -231,42 +231,6 @@ class ResponseTest {
   }
 
   #[Test]
-  public function transfer_stream_with_length() {
-    $res= new Response(new TestOutput());
-    $res->transfer(new MemoryInputStream('<h1>Test</h1>'), 'text/html', 13);
-
-    $this->assertResponse(
-      "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nContent-Length: 13\r\n\r\n".
-      "<h1>Test</h1>",
-      $res
-    );
-  }
-
-  #[Test]
-  public function transfer_stream_chunked() {
-    $res= new Response(new TestOutput());
-    $res->transfer(new MemoryInputStream('<h1>Test</h1>'), 'text/html');
-
-    $this->assertResponse(
-      "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nTransfer-Encoding: chunked\r\n\r\n".
-      "d\r\n<h1>Test</h1>\r\n0\r\n\r\n",
-      $res
-    );
-  }
-
-  #[Test]
-  public function transfer_stream_buffered() {
-    $res= new Response(new TestOutput(Buffered::class));
-    $res->transfer(new MemoryInputStream('<h1>Test</h1>'), 'text/html');
-
-    $this->assertResponse(
-      "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nContent-Length: 13\r\n\r\n".
-      "<h1>Test</h1>",
-      $res
-    );
-  }
-
-  #[Test]
   public function transmit_stream_with_length() {
     $res= new Response(new TestOutput());
     foreach ($res->transmit(new MemoryInputStream('<h1>Test</h1>'), 'text/html', 13) as $_) { }